Add ReviewList view
Add a view for reviewing, in list form, all of the mods that are supposed to be moved between characters. Add functionality for sorting this list by both the currently-assigned character and the recipient character (in order of importance), as well as marking the mods as either removed from a character or reassigned.