Add a set view to the mod assignment screen
Update the mod assignment screen to add a "sets" vs "list" toggle to the filter form. When viewing as sets, the screen will either show all the mods to unequip from a character (if sorting by currently assigned character) or all the mods to equip (if sorting by assigned character), with appropriate styling. The sets view will also unequip or reassign all mods associated with a character in a single click, rather than a button click per mod.