Gracefully handle missing tags on characters
If characters have no tags property (such as when an unknown character name is encountered and a new character is created), prevent an issue where an exception would be thrown on the ReviewList view when trying to build the filter form. Characters with missing tags will now simply not contribute to the tags list on the form.