Normalize stat values on character edit form
Update stat values to all allow a range from -100 to 100. Update the character edit form to use range sliders for all stat value inputs. Update the way that stat values are used internally to normalize them all based on the maximum value of that stat on a mod. Add a RangeInput component that displays the input along with an output field that automatically gets updated as the slider moves. Moved the option to use only 5-dot mods to the Character class from the OptimizationPlan class. OptimizationPlan now holds only stat weights.