Cleanup
Changed styling on several elements to make them better aesthetically. Changed the instructions on the character selection page to demonstrate how to use the new style of stat weights. Added the ability to double click characters to move them to the bottom of the selected characters list. Changed the "Show me the mods to move" page to sort by GP instead of character name when not sorting by selected order. Started adding in optimization target defaults.