Add buttons to reset or delete character targets
For each available target plan for a character, if that plan name is one of the defaults and the current values don't match the defaults, prefix an asterisk to the plan name to show that it is different. When editing the plan, include a "Reset target to default" button. If the plan name is not one of the defaults, include a "Delete target" button that will delete the plan name, but leave the currently selected target unchanged (resulting in a "custom" target selection)