Only render visible character edit form
Previously, when viewing the character edit modal for modifying optimization targets, both the basic and advanced forms were being rendered to the DOM, and one was being hidden with CSS based on the form state. This caused browser form validation to run on the entire form when it was submitted, which could cause issues with the hidden form. To fix this issue, the tool will now only render (and therefore only validate) the visible form.