Sort by character over mod slot in list view
In the list view, if characters shared the same GP, then their mods would end up grouped together by slot. Now, all of a character's mods are always listed with each other. If characters share the same GP, then their mods will be next to the other character's mods with the same GP, but not intermixed.