Provide a toggle to use level 15 stats or not
In each Optimization Target, the user can now specify whether to emulate upgrading mods to level 15 during optimization. If the box is checked (as it is by default for most targets), then mod primary stats will be upgraded in addition to using the level 15 set bonuses. Otherwise, stats will be used as they are, and set bonuses will match the mods' current levels.