Fix a bug that would cause unassigned mods
When multiple mod sets had the same value, it was possible for the optimizer to suggest one that left some slots unassigned. This caused an issue for pilots, where having a mod is always better than not having a mod. Now the optimizer will always suggest the set with the most slots filled.