Ensure that Characters always have Game Settings
It was possible previously to get into an invalid state during fetch if the request to the `/characters/` endpoint was cached, but a new character was found in the player's inventory. The app would fail when trying to read the character name from the null GameSettings object. Now, all Characters are given a default GameSettings object if instantiated from a default. The default GameSettings object will show the "missing" image for the character, and the name will be the same as the base ID.