BUG FIX: don't overwrite setless mod combinations
Previously, when trying to determine all the potential ways to combine useful mods into a mod set, the set made from just the individual best mod for each slot was being overwritten by the sets that combined various mod types. Now, the latter is properly concatenated to the list of candidate sets so that all are considered.