BUG FIX: Upgrade mods when calculating set value
Previously, when calculating the value of a mod set (to display only), the mods in that set were never upgraded (leveled, sliced, etc.) while trying to determine what the total stats for the set would be. Now, the mods will be upgraded pursuant to the character and target settings when calculating the value of the set.