BUG FIX: Add database upgrade code to optimizer.js
There was a bug that could occur if the browser pulled the new optimizer.js file before pulling the new main file. In that case, the optimizer.js would request the database at version 2, but not add the new required object store. This would result in an unrecoverable state for the application where it could not save character templates or player progress. By adding the upgrade code to the optimizer file, the database should always end up in the right state after upgrading its version.