BUG FIX: Honor stat restrictions with target stats
After the previous fix, it was still possible to have a primary stat restriction dropped when using a target stat if a higher-value set was found prior to the set that fulfilled all the primary stat restrictions. This update fixes that case so that even if the "worse" set that fulfills restrictions is found later, it will override the previous set that doesn't fulfill the restrictions.