BUG FIX: Handle when a mod set can't be found
Previously, in a certain case when a mod set couldn't be found for a character, the optimizer would fail with the message "can't destructure property 'modSet' of 'undefined' or 'null'". This was exacerbated by forcing complete sets for every character. Now, not only will the optimizer handle not being able to find any mod set for a character, but it will also properly back off of the forced complete sets and try again before giving up.