BUG FIX: Use discrete set bonuses w/ target stats
Previously, if a set bonus gave a partial value (10% speed on 174 base = 17.4), then that partial value was used when calculating if mods could hit a target stat. This could give incorrect responses, since the final useful value of that bonus is only the whole-number part. Now, only the whole-number part is used when calculating what value the rest of the mods need to hit.