Update the named target when applying a template
Previously, when applying a template (using append, replace, or apply targets only), the selected characters would be updated properly, but the named targets available to select on characters wouldn't be updated at all. This would show itself by sometimes not displaying a '*' next to a target name if it wasn't the same as the default. With this change, applying a template actually changes the targets available on a character, so all change indicators should display properly.