Show the review sidebar after all mods are moved
Previously, the optimizer hid the filter/sort portion of the sidebar from the review screen as soon as the last mod was moved because there was nothing else to filter or sort. With the addition of a filter for mod upgrades, that's no longer the case, so the full sidebar will always be shown.