fix: add default parameters to TargetStat constructor
In characterSettings TargetStat is initialized without an argument for the type.
Without a character to compare to it is not needed. Ex: new TargetStat('Speed', 170,174)
But the parameterlist has the type in the second position. So the 170 argument is assigned to type param, maximum argument gets assigned to the minimum param and maximum param is undefined. (Check for example Greedo's TargetStats to see this)
By adding default values to the constructor's signature this is avoided and we don't have to add arguments to the calls that wouldn't make sense.